+++ CHANGELOG.md
@@ -1,10 +1,18 @@
+# Version custom_keyboard_action
+
+## 0.0.4
+
+- add custom widget ScaffoldKeyboard
+
 ## 0.0.3
 
-* remove maxLines Widget TextFormFieldCustomKeyboard
-* add widget ScaffoldKeyboard
+- remove maxLines Widget TextFormFieldCustomKeyboard
+- add widget ScaffoldKeyboard
+
 ## 0.0.2
 
-* update maxLines Widget TextFormFieldCustomKeyboard
+- update maxLines Widget TextFormFieldCustomKeyboard
+
 ## 0.0.1
 
-* new Widget TextFormFieldCustomKeyboard
+- new Widget TextFormFieldCustomKeyboard